+++ lib/features/auth/data/datasources/auth_remote_data_source.dart
@@ -1,4 +1,4 @@
-import 'package:whatif_milkyway_app/features/auth/data/models/user_model.dart';
+import '../models/user_model.dart';
 
 abstract class AuthRemoteDataSource {
   Future<UserModel?> getCurrentUser();